+++ CHANGELOG.md
@@ -6,6 +6,8 @@
         * Armored Might now consideres armor focus bonuses 
     * Armor bonuses are now calculated more correctly (this cannot be disabled due to too many dependencies)
         * Enchantments now correctly increase armor bonuses instead of applying an unrelated bonus
+    * Mythic Armor Focus feats now properly trigger on proficencey rather than penalty type
+        * This means Mithril is treated as Heavy armor for the purposes of these feats   
 
 # Version 2.6.5
 * Fixes